Add camera

+++ srcs/intersection.c
@@ -6,7 +6,7 @@
 /*   By: jleem <jleem@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/06 10:30:27 by jleem             #+#    #+#             */
-/*   Updated: 2021/01/06 23:44:36 by jleem            ###   ########.fr       */
+/*   Updated: 2021/01/07 00:31:48 by jleem            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,8 +25,8 @@ t_hit			get_ray_intersection_from_object(t_ray *ray, t_object *obj)
 	if (obj->type == sphere && e > 0)
 	{
 		ret.object = obj;
-		ret.location = vec3_multiply(r_unit, d_flat - sqrt(e));
-		ret.distance = sqrt(vec3_square(ret.location));
+		ret.location = vec3_multiply(r_unit, d_flat - sqrtf(e));
+		ret.distance = sqrtf(vec3_square(ret.location));
 		ret.location = vec3_add(ray->origin, ret.location);
 	}
 	else
